requirements-enhanced.txt
# Enhanced Trading Bot Requirements
# Compatible with Android Termux

# Core dependencies
asyncio
aiohttp>=3.8.0
aiofiles>=0.8.0
requests>=2.28.0

# Data processing
pandas>=1.5.0
numpy>=1.21.0
python-dateutil>=2.8.0

# Web scraping (Termux compatible)
beautifulsoup4>=4.11.0
lxml>=4.9.0
html5lib>=1.1

# JSON and configuration
pydantic>=1.10.0
jsonschema>=4.0.0

# Database
sqlite3  # Built-in Python module

# Terminal UI (optional but recommended)
rich>=12.0.0
colorama>=0.4.0

# Crypto/Finance APIs
ccxt>=2.0.0
python-binance>=1.0.0

# Statistical analysis
scipy>=1.9.0
scikit-learn>=1.1.0

# HTTP requests with async support
httpx>=0.23.0

# Logging enhancements
structlog>=22.0.0

# Time handling
pytz>=2022.0

# Data validation
marshmallow>=3.17.0

# Configuration management
python-dotenv>=0.20.0

# WebSocket support for real-time data
websockets>=10.0

# Mathematical computations
sympy>=1.11.0

# Caching
diskcache>=5.4.0

# Progress bars and CLI improvements
click>=8.1.0
tqdm>=4.64.0

# Termux-specific optimizations
# These packages work well in Android Termux environment